feat(website): SEO round 2 — per-hero landing pages, internal linking & breadcrumbs
Create new ranking surface for the per-hero long-tail (GSC: ~33 per-hero
"<name> win rate / pick rate" queries at ~0% CTR, position 7-9, uncaptured
because only the aggregate /heroes page existed).

Per-hero pages
- Restructure /heroes into a layout (heroes.tsx = Outlet) + index
  (heroes.index.tsx, the existing page) + dynamic detail
  (heroes.$heroName.tsx), mirroring the existing chat layout pattern.
- New per-hero page: SSR win rate / pick rate / ban rate / matches from
  prefetched live data, hero image, intro copy, lazy matchup table, and
  internal links back to /heroes and /leaderboard. Dynamic seo() with
  per-hero title/description/canonical + Dataset & BreadcrumbList JSON-LD.
- hero-slug.ts: name<->slug helper (heroSlug / findHeroBySlug); invalid
  slugs 404 via notFound().

Internal linking + breadcrumbs
- HeroName gains an opt-in linkToDetail prop rendering a crawlable anchor to
  /heroes/<slug>; enabled in HeroScoreboardTable and HeroStatsTable so every
  hero row links to its detail page (also drives crawlLinks prerendering).
- New Breadcrumb primitive (visible, crawlable) on the detail pages.

Sitemap
- sitemap.ts fetches playable heroes (failure-safe) and emits /heroes/<slug>
  entries; sitemap route made async to await the builder.

Verified: pnpm typecheck, lint (no new warnings), build — 38 hero pages
prerender via crawlLinks — and curl against the built worker confirms real
SSR win-rate numbers, H1s, canonicals, breadcrumbs, JSON-LD, and 38 hero
URLs in sitemap.xml.
