+++ website/src/routes/sitemap[.]xml.ts
@@ -5,8 +5,8 @@ import { buildSitemapXml } from "~/lib/sitemap";
 export const Route = createFileRoute("/sitemap.xml")({
   server: {
     handlers: {
-      GET: () => {
-        return new Response(buildSitemapXml(), {
+      GET: async () => {
+        return new Response(await buildSitemapXml(), {
           headers: {
             "Content-Type": "application/xml; charset=utf-8",
             "Cache-Control": "public, max-age=3600",
